Dual Full Screen Layout
This is my screens resolution so the numbers might not be the same for everyone. What this does is make it look similar to the full single screen you already have but halves it so both screen are present instead of only one. I don't really know how to do it so it can full screen to someone else's custom screen size so this is only an example. if you need me to explain why the numbers are like that I'll gladly do so.